+++ app/scripts/api/snapshot/Transporter.js
@@ -89,6 +89,9 @@ define(function(require) {
         /* can be used to keep track of the success and error callbacks for a given message */
         this.messageCallbacks = {};
 
+        /* stored callbacks for registerLocalDataListener */
+        this.localDataChangedCallbacks = {};
+
         /*
          * Gets/SetsRequest callbacks
          * simId -> { key -> { onSucess -> function, onError -> function } }
@@ -112,6 +115,8 @@ define(function(require) {
          * Helper to route messages to appropriate handlers
          */
         this.capiMessageHandler = function(message) {
+            if (!message.handshake) { return; }
+
             switch (message.type) {
                 case SimCapiMessage.TYPES.HANDSHAKE_RESPONSE:
                     handleHandshakeResponse(message);
@@ -148,6 +153,9 @@ define(function(require) {
                     break;
                 case SimCapiMessage.TYPES.ALLOW_INTERNAL_ACCESS:
                     setDomainToShortform();
+                    break;
+                case SimCapiMessage.TYPES.REGISTERED_LOCAL_DATA_CHANGED:
+                    handleLocalDataChange(message);
                     break;
             }
         };
@@ -234,7 +242,6 @@ define(function(require) {
         };
 
         /*
-         *   @since 0.5
          *   Handles the get data message
          */
         var handleGetDataResponse = function(message) {
@@ -248,32 +255,43 @@ define(function(require) {
                 } else if (message.values.responseType === 'error') {
                     getRequests[message.values.simId][message.values.key].onError(message.values.error);
                 }
+
+                var nextQueuedRequest = getRequests[message.values.simId][message.values.key].inQueue;
+
                 delete getRequests[message.values.simId][message.values.key];
+
+                if (nextQueuedRequest) {
+                    self.getDataRequest(message.values.simId, message.values.key, nextQueuedRequest.onSuccess, nextQueuedRequest.onError);
+                }
             }
         };
 
         /*
-         *   @since 0.5
          *   Handles the set data message
          */
         var handleSetDataResponse = function(message) {
             if (message.handshake.authToken === handshake.authToken) {
-                var callbacks = setRequests[message.values.simId][message.values.key];
-                delete setRequests[message.values.simId][message.values.key];
                 if (message.values.responseType === 'success') {
-                    callbacks.onSuccess({
+                    setRequests[message.values.simId][message.values.key].onSuccess({
                         key: message.values.key,
                         value: message.values.value
                     });
                 } else if (message.values.responseType === 'error') {
-                    callbacks.onError(message.values.error);
+                    setRequests[message.values.simId][message.values.key].onError(message.values.error);
+                }
+
+                var nextQueuedRequest = setRequests[message.values.simId][message.values.key].inQueue;
+
+                delete setRequests[message.values.simId][message.values.key];
+
+                if (nextQueuedRequest) {
+                    self.setDataRequest(message.values.simId, message.values.key, nextQueuedRequest.value, nextQueuedRequest.onSuccess, nextQueuedRequest.onError, nextQueuedRequest.options);
                 }
             }
         };
 
 
         /*
-         * @since 0.5
          * Sends the GET_DATA Request
          */
         this.getDataRequest = function(simId, key, onSuccess, onError) {
@@ -297,13 +315,15 @@ define(function(require) {
                 }
             });
 
-            if (!getRequests[simId]) {
-                getRequests[simId] = {};
-            }
+            getRequests[simId] = getRequests[simId] || {};
 
             if (getRequests[simId][key]) {
-                //indicating that there is the same request in progress
-                return new Error('Get data request of ' + key + ' already in progress.');
+                getRequests[simId][key].inQueue = {
+                    onSuccess: onSuccess,
+                    onError: onError
+                };
+
+                return false;
             }
 
             getRequests[simId][key] = {
@@ -322,10 +342,9 @@ define(function(require) {
         };
 
         /*
-         * @since 0.5
-         * Sends the GET_DATA Request
+         * Sends the SET_DATA Request
          */
-        this.setDataRequest = function(simId, key, value, onSuccess, onError) {
+        this.setDataRequest = function(simId, key, value, onSuccess, onError, options) {
 
             check(simId).isString();
             check(key).isString();
@@ -346,16 +365,21 @@ define(function(require) {
                     key: key,
                     value: value,
                     simId: simId
-                }
+                },
+                options: options
             });
 
-            if (!setRequests[simId]) {
-                setRequests[simId] = {};
-            }
+            setRequests[simId] = setRequests[simId] || {};
 
             if (setRequests[simId][key]) {
-                //indicating that there is the same request in progress
-                return new Error('Set data request of ' + key + ' already in progress.');
+                setRequests[simId][key].inQueue = {
+                    value: value,
+                    onSuccess: onSuccess,
+                    onError: onError,
+                    options: options
+                };
+
+                return false;
             }
 
             setRequests[simId][key] = {
@@ -606,6 +630,40 @@ define(function(require) {
             self.sendMessage(message);
         };
 
+        var handleLocalDataChange = function(message) {
+            if (self.localDataChangedCallbacks[message.values.simId] && self.localDataChangedCallbacks[message.values.simId][message.values.key]) {
+                self.localDataChangedCallbacks[message.values.simId][message.values.key](message.values.value);
+            }
+        };
+
+        function unregisterLocalDataListener(simId, key) {
+            delete self.localDataChangedCallbacks[simId][key];
+        }
+
+        /*
+         * Register the sim to be notified when local data changes
+         */
+        this.registerLocalDataListener = function(simId, key, callback) {
+            check(simId).isString();
+            check(key).isString();
+
+            var message = new SimCapiMessage({
+                type: SimCapiMessage.TYPES.REGISTER_LOCAL_DATA_CHANGE_LISTENER,
+                handshake: this.getHandshake(),
+                values: {
+                    key: key,
+                    simId: simId
+                }
+            });
+
+            self.localDataChangedCallbacks[simId] = self.localDataChangedCallbacks[simId] || {};
+            self.localDataChangedCallbacks[simId][key] = callback;
+
+            self.sendMessage(message);
+
+            return unregisterLocalDataListener.bind(this, simId, key);
+        };
+
         /*
          * Send a VALUE_CHANGE message to the viewer with a dump of the model.
          */
@@ -730,13 +788,16 @@ define(function(require) {
 
         // handler for postMessages received from the viewer
         var messageEventHandler = function(event) {
+            var message;
             try {
-                var message = JSON.parse(event.data);
-                self.capiMessageHandler(message);
+                message = JSON.parse(event.data);
             } catch (e) {
                 window.console.log(e);
             }
 
+            if (message) {
+                self.capiMessageHandler(message);
+            }
         };
 
         // we have to wait until the dom is ready to attach anything or sometimes the js files